pymg-optimizer/pyproject.toml
[project]
name = "pymg-optimizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "logging>=0.4.9.6",
    "pillow>=11.3.0",
]
